alter blyrics-pulse to blyrics-shimmer in blyrics.css
tried to add support for original's ui text color variable instead of custom defining my preference, white.

Need to check if there does remain any scaffolding for old pulse animation anywhere which might be no longer needed/could create interference.

Need to ascertain if this is the most cpu-efficient way to induce this wave effect. My implementation relies on css only as that's only part of design what i'm slightly familiar with. This could be evaluated as my bias in this implementation.